Add a .gitattributes file to cause git-bigfile to be used for our big files.
It might be nice to have *every* file in this repo controlled by
git-bigfile, but I think it's less necessary for the lite .json files,
which change infrequently.  (Also, they are written with
sort_keys=True, so diffs will be small-ish.)

Test Plan:
Fingers crossed.  We'll see when I commit the first set of pofiles!