Add a stats directory & set it up to use bigfile.
Summary:
In https://phabricator.khanacademy.org/D21489 we expect there to be a
stats directory already made to save the stats file to.  These will be
big zips, so we add them to bigfile as well.

I added an empty es.zip file to begin with, in order to be able to add the directory at all.  This should be overwritten after the next download of es.

Test Plan: cross-fingers

Reviewers: chris, jlfwong, csilvers

Reviewed By: csilvers

Subscribers: kevinb, csilvers

Differential Revision: https://phabricator.khanacademy.org/D21576